framing: widen shared Message Length to 12 bits (common framing 0.2.0)
The shared 4-byte application message header capped a message at 255 bytes,
which capped market-by-price at N = 5 levels per side. Ten levels per side is
the requirement, so widen the length field family-wide rather than splitting
BookDepth into one-sided messages.

Message Length becomes a 12-bit value: the u8 at offset 1 supplies bits 0-7 and
the low nibble of byte 3 supplies bits 8-11, with byte 3's high nibble reserved.
Flags narrows from u16 to u8 at offset 2.

The layout is chosen so that the widening is byte-identical to 0.1.x for every
message of 255 bytes or fewer: byte 3 is zero for such messages, which is what a
0.1.x encoder already wrote as the high byte of a Flags value that only ever set
bit 0, and a 0.1.x decoder reading Flags as a u16 sees the extension nibble in
bits 8-11 that it was already required to ignore. No existing publisher or
subscriber changes, and no sibling feed's messages change on the wire.

A contiguous u16 length at offset 1 would have moved Flags and broken every
existing implementation for no functional gain.

Schema Version gates the one case old decoders cannot handle: 1 means no message
on the channel exceeds 255 bytes, 2 declares that the extension nibble may be
used, and an 8-bit-length decoder MUST discard version-2 frames rather than
mis-walk them. Top-of-book, market-by-order, midpoint, perp-stats, order-intent
and the reference-data supplement all keep Schema Version 1: every message type
they define is 148 bytes or fewer.

Top-of-book is the canonical definition (perp-stats already deferred its header
there); the other specs restate it and are marked as not permitted to diverge.

tools/conformance: assemble the 12-bit length in the frame walk, read Flags as
u8, accept Schema Version 1 and 2, and flag a length extension used under a
version-1 frame as a publisher self-inconsistency.

+++ tools/conformance/engine/tier1.go
@@ -20,7 +20,11 @@ import (
 // expectedMsgLen returns the canonical wire length (including the 4-byte
 // header) for message types that have a fixed size in the current spec version.
 // Returns 0 for types that are not defined in a given feed (caller must check).
-func expectedMsgLen(feed core.Feed, typ uint8) uint8 {
+// expectedMsgLen returns the mandated byte length of a message type on a feed, or
+// 0 when the type carries no mandated length there. The return is uint16 because
+// common framing 0.2.0 allows a message length above 255; no type on the feeds
+// decoded here reaches that, but the comparison has to be width-correct.
+func expectedMsgLen(feed core.Feed, typ uint8) uint16 {
 	switch typ {
 	case wire.TypeHeartbeat:
 		return 16
